feat/checkpoint: optimistic update, regression: can't put more than one accessory
